requirements.txt
# Django framework
Django==5.1.2

# Environment variable management
django-environ==0.9.0

# PostgreSQL adapter for Django
psycopg2-binary==2.9.7

# Production server for Django
gunicorn==20.1.0

whitenoise==6.4.0 

django-widget-tweaks==1.4.12

reportlab==4.0.4  

django-ckeditor

